[ty] Separate script and uv modules from project metadata
