+++ crates/ty_project/src/script.rs
