+++ crates/ty_project/src/uv.rs
@@ -1,0 +1,5 @@
+//! Integrates uv with project discovery.
+
+pub(crate) use metadata::UvWorkspace;
+
+mod metadata;
