+++ crates/ty_project/src/uv/metadata.rs
@@ -7,17 +7,17 @@ use serde::Deserialize;
 use thiserror::Error;
 use ty_static::EnvVars;
 
-use super::python_version::SupportedPythonVersion;
+use crate::metadata::python_version::SupportedPythonVersion;
 
 #[derive(Debug, Clone, PartialEq, Eq, get_size2::GetSize)]
-pub(super) struct UvWorkspace {
+pub(crate) struct UvWorkspace {
     root: SystemPathBuf,
     environment: Option<SystemPathBuf>,
     python_version: Option<RangedValue<SupportedPythonVersion>>,
 }
 
 impl UvWorkspace {
-    pub(super) fn discover(
+    pub(crate) fn discover(
         path: &SystemPath,
         system: &dyn System,
     ) -> Result<Self, UvWorkspaceError> {
@@ -50,7 +50,7 @@ impl UvWorkspace {
         Self::from_metadata(&output.stdout, system)
     }
 
-    pub(super) fn from_metadata(
+    pub(crate) fn from_metadata(
         metadata: &[u8],
         system: &dyn System,
     ) -> Result<Self, UvWorkspaceError> {
@@ -78,15 +78,15 @@ impl UvWorkspace {
         })
     }
 
-    pub(super) fn root(&self) -> &SystemPath {
+    pub(crate) fn root(&self) -> &SystemPath {
         &self.root
     }
 
-    pub(super) fn environment(&self) -> Option<&SystemPath> {
+    pub(crate) fn environment(&self) -> Option<&SystemPath> {
         self.environment.as_deref()
     }
 
-    pub(super) fn python_version(&self) -> Option<&RangedValue<SupportedPythonVersion>> {
+    pub(crate) fn python_version(&self) -> Option<&RangedValue<SupportedPythonVersion>> {
         self.python_version.as_ref()
     }
 }
@@ -129,7 +129,7 @@ fn existing_directory(
 }
 
 #[derive(Debug, Error)]
-pub(super) enum UvWorkspaceError {
+pub(crate) enum UvWorkspaceError {
     #[error("Failed to invoke `uv workspace metadata`: {0}")]
     Invocation(#[source] std::io::Error),
